Split original TenantFileSystemStorageTestCase to a separate file
because I'm going to make changes to it which will result in a
somewhat bigger file. This is also expected to grow over time.
