+++ django_tenants/tests/__init__.py
@@ -2,3 +2,4 @@ from .test_routes import *
 from .test_tenants import *
 from .test_cache import *
 from .test_static_files import *
+from .test_filesystem_storage import *
